fix jenkins pb [HPCTM-32] and [HPCTM-41]
int -> size_t

+++ test/regression/channels/Na.cpp
@@ -21,7 +21,7 @@ namespace Na{
     template<class T>
     struct channel{
         typedef T value_type;
-        static const int value_size = 11;
+        static const size_t value_size = 11;
     };
 
     template<class T>
